Move populate_links/1 into Membrane.ParentSpec. Update tests accordingly. Add test chcecking the populate_links/1. Mentions populate_links/1 in ParentSpec documentation
