+++ lib/membrane/parent_spec.ex
@@ -64,6 +64,13 @@ defmodule Membrane.ParentSpec do
         |> to(:another_element)
       ]
 
+   You can also use `populate_links/1` in order to link subsequent children using default pads
+   (linking `:input` to `:output` of previous element). That might be especially helpful when creating
+   testing pipelines.
+
+      children = [source: Some.Source, filter: Some.Filter, sink: Some.Sink]
+      links = populate_links(children)
+
   ### Bins
 
   For bins boundaries, there are special links allowed. The user should define links
@@ -455,6 +462,25 @@ defmodule Membrane.ParentSpec do
     end
     |> LinkBuilder.update(:to_pad, to_pad: pad, to_pad_props: %{})
     |> to({Membrane.Bin, :itself})
+  end
+
+  @doc """
+  Links subsequent children using default pads (linking `:input` to `:output` of
+  previous element).
+
+  ## Example
+
+      Membrane.ParentSpec.populate_links([el1: MembraneElement1, el2: MembraneElement2])
+  """
+  @spec populate_links(children :: ParentSpec.children_spec_t()) :: ParentSpec.links_spec_t()
+  def populate_links(children) when length(children) < 2 do
+    []
+  end
+
+  def populate_links(children) when is_list(children) do
+    [h | t] = children |> Keyword.keys()
+    links = t |> Enum.reduce(link(h), &to(&2, &1))
+    [links]
   end
 
   defp validate_pad_name(pad) when Pad.is_pad_name(pad) or Pad.is_pad_ref(pad) do
