+++ test/membrane/element_test.exs
@@ -67,7 +67,7 @@ defmodule Membrane.ElementTest do
       Testing.Pipeline.start_link(
         mode: :default,
         children: children,
-        links: Testing.Pipeline.populate_links(children)
+        links: Membrane.ParentSpec.populate_links(children)
       )
 
     on_exit(fn ->
